+++ burner/extension/src/lib/inpage-next/message.ts
@@ -1,4 +1,4 @@
-import { account } from "./account";
+import { account, SESSION_LOADED_EVENT } from "./account";
 
 const uuid = "589c80c1eb85413d";
 
@@ -13,12 +13,16 @@ export const request = (msg: any) => {
   const iframe = burner.querySelector<HTMLIFrameElement>("#iframe");
   if (!iframe) {
     throw new Error("iframe not found");
+    return;
   }
   console.log("sending now...", { ...msg, uuid });
   return iframe.contentWindow?.postMessage({ ...msg, uuid }, "*");
 };
 
-export const waitForMessage = (type: any, timeout: any) => {
+export const waitForMessage = (
+  type: string,
+  timeout: number
+): Promise<string | null> => {
   return new Promise((resolve, reject) => {
     const pid = setTimeout(() => reject(new Error("timeout")), timeout);
     const handler = (event: MessageEvent) => {
@@ -34,9 +38,20 @@ export const waitForMessage = (type: any, timeout: any) => {
 
 export const extensionEventHandler = (messageEvent: MessageEvent) => {
   if (messageEvent?.data?.uuid === uuid) {
-    account._events[messageEvent.data.type].forEach((fn: any) => {
-      console.log(`event ${messageEvent.data.type} received in extension`);
-      fn(messageEvent.data.data);
-    });
+    if (!account) {
+      return;
+    }
+    if (messageEvent?.data?.type === SESSION_LOADED_EVENT) {
+      account._events.session_loaded.forEach((fn: (data: string) => void) => {
+        console.log(`event ${messageEvent.data.type} received in extension`);
+        if (
+          !messageEvent.data.data &&
+          typeof messageEvent.data.data !== "string"
+        ) {
+          return;
+        }
+        fn(messageEvent.data.data);
+      });
+    }
   }
 };
