+++ burner/extension/src/lib/inpage-next/window.ts
