+++ burner/extension/src/main.ts
@@ -1,8 +1,3 @@
-import {wallet} from './index'
-
-document.querySelector<HTMLDivElement>("#app")!.innerHTML = `
-<div> 
-    <div class="card" id="starknetburner" /></div>
-</div>`
+import { wallet } from "./index";
 
 wallet();
