+++ burner/extension/tsconfig.json
@@ -1,7 +1,6 @@
 {
     "compilerOptions": {
       "target": "ESNext",
-      "allowJs": true,
       "useDefineForClassFields": true,
       "module": "ESNext",
       "lib": ["ESNext", "DOM"],
